requirements.txt
django
django-rest-knox
djangorestframework
django-cors-headers
djoser
psycopg2
uwsgi
requests
gunicorn